fix: update barrel exports to match renamed files
SegmentControl → SegmentedControl, usePageOverLay → usePageOverlay
