+++ components/index.ts
@@ -18,10 +18,10 @@ export * from './SearchAdvancedOptions';
 export * from './searchColorLegend';
 export * from './SearchEmptyState';
 export * from './SearchInput';
-export * from './searchResultItem';
+export * from './SearchResultItem';
 export * from './SearchSkeleton';
 export * from './SearchSkeletonBar';
-export * from './SegmentControl';
+export * from './SegmentedControl';
 export * from './SegmentedControlWithDisabled';
 export * from './SelectRiwaya';
 export * from './Seo';
